visualizer/requirements.txt
plotly
grpcio-tools
flask